+++ src/zplProvider.ts
@@ -25,6 +25,12 @@ export class ZPLEditorProvider implements vscode.CustomTextEditorProvider {
     webviewPanel: vscode.WebviewPanel,
     token: vscode.CancellationToken
   ): void | Thenable<void> {
+    vscode.workspace
+      .createFileSystemWatcher(document.uri.fsPath)
+      .onDidChange(() => {
+        this.updateWebview(document, webviewPanel);
+      });
+
     webviewPanel.webview.options = {
       enableScripts: true,
     };
@@ -126,15 +132,19 @@ export class ZPLEditorProvider implements vscode.CustomTextEditorProvider {
     document: vscode.TextDocument,
     webviewPanel: vscode.WebviewPanel
   ): Promise<void> {
-    return zplToPng(document.getText(), this.dpmm).then(
-      (label: string | void) => {
+    return zplToPng(document.getText(), this.dpmm)
+      .then((label: string | void) => {
         if (label) {
           webviewPanel.webview.html = this.getHtmlForWebview(
             webviewPanel.webview,
             label
           );
         }
-      }
-    );
+      })
+      .catch((err: Error) => {
+        vscode.window.showErrorMessage(
+          `Error rendering ZPL preview: ${err.message}`
+        );
+      });
   }
 }